requirements.txt
blessed==1.17.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
future==0.18.2
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
h2o==3.30.0.1
idna==2.9
numpy==1.18.2
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
tabulate==0.8.7
typesentry==0.2.7
urllib3==1.25.8
wcwidth==0.1.9
